src:{cmd,lib}: add server enroll code
The current implementation doesn't allow for a livesystem to carry out
the exchange flow. Add a small flask instance that leans on the
lib/redactor module to host this long-living process. We use HTTP for
simplicity.

Also, to simplify things, I added a target to run the server under make.
I updated some dependencies to ensure things weren't broken