src:Redactor: reorganize model and database code
The storage/retrieval functions under redactor should also be in the
database module. The database should be the only one that is aware on
how storage and retrieval is done. The rest of the code should
exclusively focus on the rest of the functionality.

Similarly, a models module should contain the interface the SMT, as well
as issuing new trees on demand. The hierarchy is:

sboms -> produces flattened SBOM objects
models -> Produces trees from database or sbom objects
database -> persists the models into the database