Reorganize modules (#10)
* remove generated and unused files

* lib: move library code into petra module

This should simplify the separation between commands and librarified
code (i.e., that which can be integrated into other software). This
should also help us keep things neat/clean.

* Redaction: Move database to separate module

Having a separate module for the database operations will let use reason
about the Redaction script as a standalone application that *only* uses
the library code. This is because this script should exclusively focus
on interfacing the library with the user code.

* src:Redactor: move SBOM handling to library code

Same rationale as database backend. We can re-use this code where
necessary, and simplifies the script, which should exclusively focused
on presenting this functionality to users.

* src:Redactor: reorganize model and database code

The storage/retrieval functions under redactor should also be in the
database module. The database should be the only one that is aware on
how storage and retrieval is done. The rest of the code should
exclusively focus on the rest of the functionality.

Similarly, a models module should contain the interface the SMT, as well
as issuing new trees on demand. The hierarchy is:

sboms -> produces flattened SBOM objects
models -> Produces trees from database or sbom objects
database -> persists the models into the database

* src:Redactor: mode crypto in to separate module

Same rationale as the above. A crypto-focused module should allow us to
manage the key generation/management as well as encrypting data on
demand.