petra:{models,sbom} add format agnostic models
Building the models using lib4sbom may result in missing some fields on
the parsed SBOM. Instead, we can rely on a format-agnosting tree walk on
any tree representation (i.e., anything json, xml or yaml) with a series
of base objects. This is ideal, at least for the JSON case, as SBOMs in
this format really just have 3 types:

1. key, value pairs as strings
2. Named objects (i.e., dicts)
3. Collections of anonymous objects (i.e., lists with dicts inside)

Write a very simple parser to handle this edgecase to build an SBOM
tree. Also, simplify the visitor construction to help visualize the tree
parsing routine on a selected SBOM (hand-crafted by yours truly) and
serve as the foundation for a future testing.

With these building blocks, we are likely able to implement minimal
visitor objects that can be applied on this base tree object. Also,
using object composition we may be able to keep a stateful "super tree"
structure in future commits.